+++ src/main/kotlin/com/atlassian/performance/tools/jiraperformancetests/api/OnPremisePerformanceTest.kt
@@ -1,6 +1,7 @@
 package com.atlassian.performance.tools.jiraperformancetests.api
 
 import com.atlassian.performance.tools.infrastructure.api.virtualusers.LocalVirtualUsers
+import com.atlassian.performance.tools.jiraactions.api.scenario.Scenario
 import com.atlassian.performance.tools.report.api.HistoricalCohortsReporter
 import com.atlassian.performance.tools.virtualusers.api.VirtualUserLoad
 import com.atlassian.performance.tools.virtualusers.api.VirtualUserOptions
@@ -14,30 +15,30 @@ import java.time.Duration
 class OnPremisePerformanceTest(
     private val jira: URI
 ) {
-    var adminLogin: String? = null
-    var adminPassword: String? = null
-    var virtualUsers: Int = 10
+    private companion object {
+        private val DEFAULTS = VirtualUserOptions()
+    }
+
+    var adminLogin: String = DEFAULTS.adminLogin
+    var adminPassword: String = DEFAULTS.adminPassword
+    var scenario: Class<out Scenario> = DEFAULTS.scenario
+    var virtualUsers: Int = DEFAULTS.virtualUserLoad.virtualUsers
     var testDuration: Duration = Duration.ofMinutes(30)
     var workspace: RootWorkspace = RootWorkspace()
     private val testName = "Benchmark my Jira"
     private val cohortName = "my-jira"
 
     fun run() {
-        var options = VirtualUserOptions(
+        val options = VirtualUserOptions(
+            adminLogin = adminLogin,
+            adminPassword = adminPassword,
+            scenario = scenario,
             jiraAddress = jira,
             virtualUserLoad = VirtualUserLoad(
                 virtualUsers = virtualUsers,
                 flat = testDuration
             )
         )
-
-        if (adminLogin != null) {
-            options = options.copy(adminLogin = adminLogin!!)
-        }
-
-        if (adminPassword != null) {
-            options = options.copy(adminPassword = adminPassword!!)
-        }
 
         BtfJiraPerformanceMeter().run(
             virtualUsers = LocalVirtualUsers(workspace = workspace.currentTask.directory),