+++ CHANGELOG.md
@@ -18,6 +18,10 @@ The API consists of all public Java types from `com.atlassian.performance.tools.
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/jira-performance-tests/branches/compare/master%0Drelease-1.1.0
 
+### Added
+- Add an option to test on premise Jira with custom test scenario which resolves 
+[JPREF-84](https://ecosystem.atlassian.net/browse/JPERF-84)
+
 ## [1.1.0]
 [1.1.0]: https://bitbucket.org/atlassian/jira-performance-tests/branches/compare/release-1.1.0%0Drelease-1.0.0