+++ CHANGELOG.md
@@ -16,7 +16,10 @@ The API consists of all public Java types from `com.atlassian.performance.tools.
 [behavioral compatibility]: http://cr.openjdk.java.net/~darcy/OpenJdkDevGuide/OpenJdkDevelopersGuide.v0.777.html#behavioral_compatibility
 
 ## [Unreleased]
-[Unreleased]: https://bitbucket.org/atlassian/jira-performance-tests/branches/compare/master%0Drelease-1.1.0
+[Unreleased]: https://bitbucket.org/atlassian/jira-performance-tests/branches/compare/master%0Drelease-1.2.0
+
+## [1.2.0]
+[1.2.0]: https://bitbucket.org/atlassian/jira-performance-tests/branches/compare/release-1.2.0%0Drelease-1.1.0
 
 ### Added
 - Add an option to test on premise Jira with custom test scenario, which resolves [JPERF-84].